added the 2-largest_number.c file

+++ 0x03-debugging/2-largest_number.c
@@ -8,9 +8,9 @@
  * Return: largest number
  */
 
-long int largest_number(long int a, long int b, long int c)
+int largest_number(int a, int b, int c)
 {
-	 long int largest;
+	 int largest;
 
 	 if ((a > b) && (a > c))
 	 {
